注解过滤器工厂
