+++ boots/simboot-core-annotation/src/main/kotlin/love/forte/simboot/annotation/AnnotationEventFilterFactory.kt
@@ -1,15 +1,41 @@
 package love.forte.simboot.annotation
 
 import love.forte.simbot.Api4J
+import love.forte.simbot.MutableAttributeMap
 import love.forte.simbot.event.*
-import javax.inject.Singleton
 
 /**
+ * 应用于 [@Filter][love.forte.simboot.annotation.Filter] 注解上的
+ * [Filter.by][love.forte.simboot.annotation.Filter.by] 参数，用于
+ * 通过参数构建一个当前 `Filter` 对应的过滤器实例。
  *
+ * ```kotlin
+ * @Filter(by = FooAnnotationEventFilterFactory::class)
+ * suspend fun Event.onEvent() { ... }
+ * ```
+ *
+ * @author ForteScarlet
  */
 public interface AnnotationEventFilterFactory {
-    // TODO
+    
+    
+    /**
+     * 通过提供的监听函数和过滤器注解参数来解析并得到一个 [EventFilter] 实例。
+     *
+     * 如果需要跳过本次解析，可以直接返回一个 null。
+     *
+     * @see Filter.by
+     *
+     */
+    public fun resolveFilter(
+        listener: EventListener,
+        listenerAttributes: MutableAttributeMap,
+        filter: Filter,
+        filters: Filters,
+    ): EventFilter?
+    
 }
+
 
 /**
  * 应用于 [love.forte.simboot.annotation.Filter] 注解上用来直接处理对应注解的函数。
@@ -23,62 +49,9 @@ public interface AnnotationEventFilterFactory {
 @Suppress("KDocUnresolvedReference")
 public interface AnnotationEventFilter : EventFilter {
     
-    
-    /**
-     * 在此实例被获取/构建之后、应用之前被调用, 用于通过监听函数实例和对应的过滤器注解参数来对当前过滤器进行初始化。
-     *
-     * ## 使用
-     * [AnnotationEventFilter] 应用于 [love.forte.simboot.annotation.Filter.by] 中，具体使用描述参考其文档注释。
-     * ```kotlin
-     * @Filter(by = FooAnnotationEventFilter::class)
-     * suspend fun Event.onEvent() { ... }
-     * ```
-     *
-     * ## 实例获取
-     *
-     * 如果当前过滤器存在于依赖管理之下并且作为唯一实例使用，则初始化函数 [init] 可能会执行多次；
-     * 如果当前过滤器为 `object` 类型，[init] 同样可能会执行多次。
-     * ```kotlin
-     * object FooAnnotationEventFilter : AnnotationEventFilter {
-     *      // ...
-     * }
-     * class BarAnnotationEventFilter : AnnotationEventFilter {
-     *      // ...
-     * }
-     * ```
-     *
-     * 如果当前过滤器不存在于任何依赖管理中并且为普通的类，则会为每一个监听函数构建一个唯一的实例,
-     * 此时 [init] 将会只执行一次。
-     *
-     *
-     *
-     * 假如 [AnnotationEventFilter] 的实现不是一个 `object` 类型，但是希望对所有的监听函数来讲
-     * 其实例唯一，那么在使用的此类类型上标记 [Singleton]。标记了 [Singleton] 的普通实现，
-     * 会在构造一次实例后进行缓存，保证所有监听函数对应的实例唯一。
-     * 此时 [init] 将可能执行多次。
-     * ```kotlin
-     * @javax.inject.Singleton
-     * class FooAnnotationEventFilter : AnnotationEventFilter {
-     *      // ...
-     * }
-     * ```
-     *
-     * 除了交由依赖容器所管理的类型之外，所有要使用的实现类型必须保证存在一个公开无参的构造函数用于进行实例化。
-     *
-     * @return 此监听函数初始化后的整合类型
-     *
-     * @see Filter.by
-     * @see InitType
-     *
-     */
     public fun init(listener: EventListener, filter: Filter, filters: Filters)
     
     
-    /**
-     * [AnnotationEventFilter] 初始化后, 用于代表当前过滤器针对于当前监听函数的状态类型。
-     *
-     * @see AnnotationEventFilter.init
-     */
     @Deprecated("Unused")
     public enum class InitType {
         INDEPENDENT,
@@ -86,9 +59,6 @@ public interface AnnotationEventFilter : EventFilter {
     }
     
     
-    /**
-     * 过滤器的检测函数。通过 [EventProcessingContext] 来验证是否需要处理当前事件。
-     */
     override suspend fun test(context: EventListenerProcessingContext): Boolean
     
 }
